rename solver file
