+++ qutip/solver/solver_base.py
